requirements.txt
ale_py==0.7.4
beautifulsoup4==4.11.2
clip==1.0
dotmap==1.3.30
duckdb==0.6.1
eval==0.0.5
gym==0.26.2
h5py==3.8.0
ipython==8.11.0
librosa==0.10.0
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.6.0.66
pafy==0.5.5
pandas==1.5.2
Pillow==9.4.0
playsound==1.3.0
PyYAML==6.0
requests==2.28.1
sentence_transformers==2.2.2
skimage==0.0
sqlglot==10.2.6
stable_baselines3==1.6.2
streamlit==1.15.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
trainer==0.0.22
transformers==4.25.1
